microservices: Refactor document ingestion service
