+++ microservices/document-ingestion/pgvector/app/.env
@@ -1,0 +1,1 @@
+LOCAL_STORE_PREFIX = "/tmp/intelgai"
