+++ microservices/document-ingestion/pgvector/pyproject.toml
@@ -9,10 +9,13 @@ readme = "README.md"
 python = "^3.10"
 av = "13.1.0"
 beautifulsoup4 = "4.12.3"
+boto3 = "1.38.0"
 docarray = {extras = ["full"], version = "0.40.0"}
 docx2txt = "0.8"
 fastapi = "0.115.4"
+gunicorn = "^23.0.0"
 huggingface_hub = "0.26.2"
+html2text = "2024.2.26"
 langchain = "^0.3.9"
 langchain-community = "^0.3.9"
 langchain-text-splitters = ">=0.3.3,<0.4.0"
@@ -28,11 +31,13 @@ pytesseract = "0.3.13"
 python-docx = "1.1.2"
 python-multipart = ">=0.0.18"
 python-pptx = "1.0.2"
+pydantic = "^2.9.2"
+pydantic-settings = "^2.6.1"
 shortuuid = "1.0.13"
+structlog = "25.1.0"
 tiktoken = "0.8.0"
 unstructured = {extras = ["pdf", "docx"], version = ">=0.14.3"}
 uvicorn = ">=0.21.0,<0.23.0"
-html2text = "2024.2.26"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^8.3.3"
